cbor: drop openssl < 1.1 HMAC_Update() casts
missed in bf15362; fixes build on xenial (which ships openssl 1.0.2)

+++ src/cbor.c
@@ -726,8 +726,8 @@ cbor_encode_change_pin_auth(const fido_blob_t *key, const fido_blob_t *new_pin,
 
 	if ((md = EVP_sha256()) == NULL ||
 	    HMAC_Init_ex(&ctx, key->ptr, (int)key->len, md, NULL) == 0 ||
-	    HMAC_Update(&ctx, npe->ptr, (int)npe->len) == 0 ||
-	    HMAC_Update(&ctx, phe->ptr, (int)phe->len) == 0 ||
+	    HMAC_Update(&ctx, npe->ptr, npe->len) == 0 ||
+	    HMAC_Update(&ctx, phe->ptr, phe->len) == 0 ||
 	    HMAC_Final(&ctx, dgst, &dgst_len) == 0 || dgst_len != 32) {
 		fido_log_debug("%s: HMAC", __func__);
 		goto fail;
